+++ pybamm/expression_tree/functions.py
@@ -21,10 +21,19 @@ class Function(pybamm.Symbol):
     derivative : str, optional
         Which derivative to use when differentiating ("autograd" or "derivative").
         Default is "autograd".
+    differentiated_function : method, optional
+        The function which was differentiated to obtain this one. Default is None.
     **Extends:** :class:`pybamm.Symbol`
     """
 
-    def __init__(self, function, *children, name=None, derivative="autograd"):
+    def __init__(
+        self,
+        function,
+        *children,
+        name=None,
+        derivative="autograd",
+        differentiated_function=None
+    ):
 
         if name is not None:
             self.name = name
@@ -39,6 +48,7 @@ class Function(pybamm.Symbol):
 
         self.function = function
         self.derivative = derivative
+        self.differentiated_function = differentiated_function
 
         # hack to work out whether function takes any params
         # (signature doesn't work for numpy)
@@ -85,7 +95,9 @@ class Function(pybamm.Symbol):
                 # if variable appears in the function,use autograd to differentiate
                 # function, and apply chain rule
                 if variable.id in [symbol.id for symbol in child.pre_order()]:
-                    partial_derivatives[i] = self._diff(children) * child.diff(variable)
+                    partial_derivatives[i] = self._diff(children, i) * child.diff(
+                        variable
+                    )
 
             # remove None entries
             partial_derivatives = list(filter(None, partial_derivatives))
@@ -96,15 +108,31 @@ class Function(pybamm.Symbol):
 
             return derivative
 
-    def _diff(self, children):
+    def _diff(self, children, idx):
         """ See :meth:`pybamm.Symbol._diff()`. """
+        # Store differentiated function, needed in case we want to convert to CasADi
         if self.derivative == "autograd":
-            return Function(autograd.elementwise_grad(self.function), *children)
-        elif self.derivative == "derivative":
-            # keep using "derivative" as derivative
-            return pybamm.Function(
-                self.function.derivative(), *children, derivative="derivative"
+            return Function(
+                autograd.elementwise_grad(self.function, idx),
+                *children,
+                differentiated_function=self.function
             )
+        elif self.derivative == "derivative":
+            if len(children) > 1:
+                raise ValueError(
+                    """
+                    differentiation using '.derivative()' not implemented for functions
+                    with more than one child
+                    """
+                )
+            else:
+                # keep using "derivative" as derivative
+                return pybamm.Function(
+                    self.function.derivative(),
+                    *children,
+                    derivative="derivative",
+                    differentiated_function=self.function
+                )
 
     def _function_jac(self, children_jacs):
         """ Calculate the jacobian of a function. """
@@ -118,7 +146,7 @@ class Function(pybamm.Symbol):
             children = self.orphans
             for i, child in enumerate(children):
                 if not child.evaluates_to_number():
-                    jac_fun = self._diff(children) * children_jacs[i]
+                    jac_fun = self._diff(children, i) * children_jacs[i]
                     jac_fun.domain = []
                     if jacobian is None:
                         jacobian = jac_fun
@@ -175,7 +203,11 @@ class Function(pybamm.Symbol):
             A new copy of the function
         """
         return pybamm.Function(
-            self.function, *children, name=self.name, derivative=self.derivative
+            self.function,
+            *children,
+            name=self.name,
+            derivative=self.derivative,
+            differentiated_function=self.differentiated_function
         )
 
     def _function_simplify(self, simplified_children):
@@ -203,7 +235,8 @@ class Function(pybamm.Symbol):
                 self.function,
                 *simplified_children,
                 name=self.name,
-                derivative=self.derivative
+                derivative=self.derivative,
+                differentiated_function=self.differentiated_function
             )
 
 
@@ -239,7 +272,7 @@ class Cos(SpecificFunction):
     def __init__(self, child):
         super().__init__(np.cos, child)
 
-    def _diff(self, children):
+    def _diff(self, children, idx):
         """ See :meth:`pybamm.Symbol._diff()`. """
         return -Sin(children[0])
 
@@ -255,7 +288,7 @@ class Cosh(SpecificFunction):
     def __init__(self, child):
         super().__init__(np.cosh, child)
 
-    def _diff(self, children):
+    def _diff(self, children, idx):
         """ See :meth:`pybamm.Function._diff()`. """
         return Sinh(children[0])
 
@@ -271,7 +304,7 @@ class Exponential(SpecificFunction):
     def __init__(self, child):
         super().__init__(np.exp, child)
 
-    def _diff(self, children):
+    def _diff(self, children, idx):
         """ See :meth:`pybamm.Function._diff()`. """
         return Exponential(children[0])
 
@@ -287,7 +320,7 @@ class Log(SpecificFunction):
     def __init__(self, child):
         super().__init__(np.log, child)
 
-    def _diff(self, children):
+    def _diff(self, children, idx):
         """ See :meth:`pybamm.Function._diff()`. """
         return 1 / children[0]
 
@@ -313,7 +346,7 @@ class Sin(SpecificFunction):
     def __init__(self, child):
         super().__init__(np.sin, child)
 
-    def _diff(self, children):
+    def _diff(self, children, idx):
         """ See :meth:`pybamm.Function._diff()`. """
         return Cos(children[0])
 
@@ -329,7 +362,7 @@ class Sinh(SpecificFunction):
     def __init__(self, child):
         super().__init__(np.sinh, child)
 
-    def _diff(self, children):
+    def _diff(self, children, idx):
         """ See :meth:`pybamm.Function._diff()`. """
         return Cosh(children[0])