+++ pybamm/solvers/ode_solver.py
@@ -1,6 +1,7 @@
 #
 # Base solver class
 #
+import casadi
 import pybamm
 import numpy as np
 
@@ -102,13 +103,13 @@ class OdeSolver(pybamm.BaseSolver):
                 pybamm.logger.info("Simplifying jacobian")
                 jac_rhs = simp.simplify(jac_rhs)
 
-            if model.use_to_python:
+            if model.convert_to_format == "python":
                 pybamm.logger.info("Converting jacobian to python")
                 jac_rhs = pybamm.EvaluatorPython(jac_rhs)
         else:
             jac_rhs = None
 
-        if model.use_to_python:
+        if model.convert_to_format == "python":
             pybamm.logger.info("Converting RHS to python")
             concatenated_rhs = pybamm.EvaluatorPython(concatenated_rhs)
             pybamm.logger.info("Converting events to python")
@@ -147,6 +148,83 @@ class OdeSolver(pybamm.BaseSolver):
         self.y0 = y0
         self.dydt = dydt
         self.events = events
+        self.event_funs = event_funs
+        self.jacobian = jacobian
+
+    def set_up_casadi(self, model):
+        """Convert model to casadi format and use their inbuilt functionalities.
+
+        Parameters
+        ----------
+        model : :class:`pybamm.BaseModel`
+            The model whose solution to calculate. Must have attributes rhs and
+            initial_conditions
+
+        Raises
+        ------
+        :class:`pybamm.SolverError`
+            If the model contains any algebraic equations (in which case a DAE solver
+            should be used instead)
+
+        """
+        # Check for algebraic equations
+        if len(model.algebraic) > 0:
+            raise pybamm.SolverError(
+                """Cannot use ODE solver to solve model with DAEs"""
+            )
+
+        y0 = model.concatenated_initial_conditions[:, 0]
+
+        t_casadi = casadi.SX.sym("t")
+        y_casadi = casadi.SX.sym("y", len(y0))
+        pybamm.logger.info("Converting RHS to CasADi")
+        concatenated_rhs = model.concatenated_rhs.to_casadi(t_casadi, y_casadi)
+        pybamm.logger.info("Converting events to CasADi")
+        casadi_events = {
+            name: event.to_casadi(t_casadi, y_casadi)
+            for name, event in model.events.items()
+        }
+
+        # Create function to evaluate rhs
+        concatenated_rhs_fn = casadi.Function(
+            "rhs", [t_casadi, y_casadi], [concatenated_rhs]
+        )
+
+        def dydt(t, y):
+            pybamm.logger.debug("Evaluating RHS for {} at t={}".format(model.name, t))
+            dy = concatenated_rhs_fn(t, y).full()
+            return dy[:, 0]
+
+        # Create event-dependent function to evaluate events
+        def event_fun(event):
+            casadi_event_fn = casadi.Function("event", [t_casadi, y_casadi], [event])
+
+            def eval_event(t, y):
+                return casadi_event_fn(t, y)
+
+            return eval_event
+
+        event_funs = [event_fun(event) for event in casadi_events.values()]
+
+        # Create function to evaluate jacobian
+        if model.use_jacobian:
+
+            pybamm.logger.info("Calculating jacobian")
+            casadi_jac = casadi.jacobian(concatenated_rhs, y_casadi)
+            casadi_jac_fn = casadi.Function(
+                "jacobian", [t_casadi, y_casadi], [casadi_jac]
+            )
+
+            def jacobian(t, y):
+                return casadi_jac_fn(t, y)
+
+        else:
+            jacobian = None
+
+        # Add the solver attributes
+        self.y0 = y0
+        self.dydt = dydt
+        self.events = model.events
         self.event_funs = event_funs
         self.jacobian = jacobian